requirements.txt
asgiref==3.8.1
cffi==1.17.1
cryptography==43.0.3
Django==5.0.1
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
pycparser==2.22
PyJWT==2.9.0
python-dotenv==1.0.0
pytz==2024.2
sqlparse==0.5.1
tzdata==2024.2
